Add severities to per_file rule
