+++ src/codechecker.bzl
@@ -388,6 +388,7 @@ def codechecker_test(
             options = analyze,
             skip = skip,
             config = config,
+            severities = severities,
             toolchain = toolchain,
             tags = codechecker_tags,
             **kwargs
