+++ src/per_file.bzl
@@ -107,6 +107,7 @@ def _run_code_checker(
             ctx.attr._per_file_script[DefaultInfo].files_to_run,
         ],
         arguments = [
+            "--mode=Run",
             "--codechecker",
             info.codechecker.path,
             "--commands",
@@ -202,6 +203,7 @@ def _per_file_impl(ctx):
         info = ctx.attr.toolchain[platform_common.ToolchainInfo].codecheckerinfo
     else:
         info = ctx.toolchains["//:toolchain_type"].codecheckerinfo
+
     for target in ctx.attr.targets:
         if not CcInfo in target:
             continue
@@ -230,29 +232,83 @@ def _per_file_impl(ctx):
                         sources_and_headers,
                     )
                     all_files += outputs
+
+    # Parse action: collect all plists into a directory and run
+    # CodeChecker parse to produce result.txt, result.json, HTML report
+    codechecker_files = ctx.actions.declare_directory(
+        ctx.label.name + "/parse",
+    )
+    codechecker_parse_log = ctx.actions.declare_file(
+        ctx.label.name + "/codechecker_parse.log",
+    )
+
+    # Build arguments for the parse action
+    # The data dir is where the per-file analyze actions put their plists
+    # All plists are in <name>/data/, derive path from first plist
+    #data_dir_path = plist_and_metadata_files[0].dirname if plist_and_metadata_files else ""
+
+    ctx.actions.run(
+        inputs = all_files + [config_file],
+        outputs = [codechecker_files, codechecker_parse_log],
+        executable = per_file_script,
+        tools = [
+            info.runfiles,
+            ctx.attr._per_file_script[DefaultInfo].files_to_run,
+        ],
+        arguments = [
+            "--mode",
+            "Parse",
+            "--codechecker",
+            info.codechecker.path,
+            "--data_dir",
+            codechecker_files.path,
+            "--log",
+            codechecker_parse_log.path,
+            "--config",
+            config_file.path,
+        ],
+        mnemonic = "CodeCheckerParse",
+        use_default_shell_env = True,
+        progress_message = "CodeChecker parse %s" % str(ctx.label),
+    )
+
+    all_files += [codechecker_files, codechecker_parse_log]
+
+    launcher = ctx.actions.declare_file(ctx.label.name + "_launcher.sh")
     ctx.actions.write(
-        output = ctx.outputs.test_script,
+        output = launcher,
+        content = """#!/bin/bash
+            exec {tool} --mode=Test \
+            --data_dir '{codechecker_files}' --severities '{severities}'
+            """.format(
+            tool = per_file_script.short_path,
+            codechecker_files = codechecker_files.short_path,
+            severities = " ".join(ctx.attr.severities),
+        ),
         is_executable = True,
-        content = """
-            DATA_DIR=$(dirname {dirname})
-            # ls -la $DATA_DIR/data
-            # find $DATA_DIR/data -name *.plist -exec sed -i -e "s|<string>.*execroot/codechecker_bazel/|<string>|g" {{}} \\;
-            # cat $DATA_DIR/data/test-src-lib.cc_clangsa.plist
-            echo "Running: CodeChecker parse $DATA_DIR/data"
-            $(realpath {codechecker}) parse $DATA_DIR/data
-        """.format(dirname = ctx.outputs.test_script.short_path, codechecker = info.codechecker.short_path),
     )
-    files = depset(
-        direct = all_files,
-    )
+
+    # Return test script and all required files
     run_files = [
-        ctx.outputs.test_script,
+        per_file_script,
+        launcher,
+        codechecker_files,
     ] + info.runfiles.to_list() + all_files
+    all_runfiles = ctx.runfiles(files = run_files)
+
+    # Add runfiles from the py_binary target (for common.py etc.)
+    all_runfiles = all_runfiles.merge(
+        ctx.attr._per_file_script[DefaultInfo].default_runfiles,
+    )
+
     return [
         DefaultInfo(
-            files = files,
-            runfiles = ctx.runfiles(files = run_files),
-            executable = ctx.outputs.test_script,
+            files = depset(all_files),
+            runfiles = all_runfiles,
+            executable = launcher,
+        ),
+        OutputGroupInfo(
+            codechecker_files = depset([codechecker_files]),
         ),
     ]
 
@@ -273,6 +329,10 @@ per_file_test = rule(
         "options": attr.string_list(
             default = [],
             doc = "List of CodeChecker options, e.g.: --ctu",
+        ),
+        "severities": attr.string_list(
+            default = ["HIGH"],
+            doc = "List of defect severities: HIGH, MEDIUM, LOW, STYLE etc",
         ),
         "skip": attr.string_list(
             default = [],
@@ -300,7 +360,6 @@ per_file_test = rule(
     },
     outputs = {
         "compile_commands": "%{name}/compile_commands.json",
-        "test_script": "%{name}/test_script.sh",
     },
     test = True,
     toolchains = ["//:toolchain_type"],
